Fixed issue with state tree database adapter
DBAdapter parent transaction is no longer renewed in response to the batch writer closing (this was done to initiate the tx renew event).

The consequence of batch writer renewing the tx means uncommitted operations in the parent event are lost.

Also, when the parent tx is not renewed, It means iterators do not see what the batcher is doing since they are using a tx that is isolated.

To solve this, each iterator starts on a different managed transaction, allowing them to operate independently and not be renewed whenever the parent tx is renewed/discarded