Config command now adds kit as `core.askPass` helper command
This means it will respond with signed push tokens when git requests
for a password. This helps mitigate the issue where push tokens generated
during `pre-push` hook are not used/seen by git.

The askPass helper collects already signed tokens that may have been created
by kit's hook/sign commands and returns them to git.

Note: It might pass invalid/stale tokens to kit in some cases. This is because
git does not provide enough information to nicely select a specific remote, instead
we try to find any tokens generated for a remote that has a push url that
matches the hostname requested by git (which is the only criterion we have to filter with).

We may need to do more work on the remote server side to ignore
invalid tokens signed for a reference that was not included in the push request.