Moved dht package into new net package
Given that the DHT and Tendermint p2p operate on different networks, we
are now classifying the DHT network to be an overlay network that may
interact with the Tendermint p2p network in the future.

I have now moved the DHT into the new `net` package and also moved the creation
of the libp2p host out of the DHT server. This is necessary to allow other
services and components share the same libp2p hosts. Some day when Tendermint
moves to libp2p, the `net` host will be connected to Tendermint via a circuit so that
we end up with one transport for all networks.